IT WORKSSSSSS WOOOOOOOOOOO

+++ include/component/BlockTextureComponent.h
@@ -19,7 +19,7 @@ public:
     std::shared_ptr<GLuint> m_meshVboId;
     std::shared_ptr<GLuint> m_uvVboId;
     std::shared_ptr<GLuint> m_texVboId;
-    std::vector<GLuint> m_textureIDs;
+    std::vector<GLfloat> m_textureIDs;
     static std::shared_ptr<GLuint> s_vaoID;
 
     static std::vector<std::string> s_registeredTextures;
